refectro: css를 styled-components로 변경 및 폴더구조 변경
