+++ src/Main.jsx
@@ -1,5 +1,6 @@
 import { BrowserRouter, Routes } from "react-router-dom";
 import { routes } from "./routes";
+
 function Main() {
   return (
     <BrowserRouter>
