--- src/components/Components/Card.css
@@ -1,77 +1,0 @@
-.img-container {
-  overflow: hidden;
-}
-.cards {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-  flex-wrap: wrap;
-  column-gap: 2rem;
-  row-gap: 2.5rem;
-}
-
-.card:hover .card-img {
-  transform: scale(1.3);
-  transition: all 1s;
-}
-
-.card {
-  background-color: var(--white-color);
-  position: relative;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  width: 34rem;
-  height: 33.4rem;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-}
-
-.card-img {
-  width: 34rem;
-  height: 20rem;
-  object-fit: cover;
-}
-
-.star-img {
-  position: absolute;
-  right: 1.5rem;
-  top: 1.5rem;
-}
-
-.card-info {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  width: 100%;
-  padding: 1.5rem 2rem;
-}
-
-.card-info-top {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.time-ago {
-  color: #666;
-  font-size: 13px;
-  font-weight: 400;
-}
-
-.links-description {
-  overflow: hidden;
-  color: #000;
-  text-overflow: ellipsis;
-  font-size: 1.6rem;
-  font-weight: 400;
-  line-height: 24px;
-  height: 4.9rem;
-  width: 100%;
-}
-
-.createdAt {
-  color: #333;
-  white-space: nowrap;
-  font-size: 1.4rem;
-  font-weight: 400;
-}
