--- src/components/Components/FolderButton.css
@@ -1,12 +1,0 @@
-.folder-button {
-  padding: 8px 12px;
-  border-radius: 5px;
-  border: 1px solid var(--primary-color);
-  background: var(--white-color);
-}
-
-.folder-button:hover,
-.folder-button:focus {
-  background-color: var(--primary-color);
-  color: var(--white-color);
-}
