--- src/components/Components/FolderCardList.css
@@ -1,9 +1,0 @@
-.noLink {
-  color: #000;
-  font-weight: 400;
-  font-size: 1.6rem;
-  text-align: center;
-  width: 100%;
-  padding-top: 4.1rem;
-  padding-bottom: 3.5rem;
-}
