--- src/components/Components/FolderLists.css
@@ -1,40 +1,0 @@
-.folders {
-  display: flex;
-  flex-direction: column;
-  gap: 2.4rem;
-}
-
-.folders-top {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.folder-buttons {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 1.2rem;
-  flex-wrap: wrap;
-}
-
-.folder-button {
-  height: 4rem;
-}
-
-.addIcon-img {
-  width: 1.6rem;
-  height: 1.6rem;
-}
-
-.folder-name {
-  font-size: 2.4rem;
-  font-weight: 600;
-}
-
-.name-and-option {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  row-gap: 1.2rem;
-}
