--- src/components/Components/Kebab.css
@@ -1,37 +1,0 @@
-.kebab-img {
-  position: relative;
-  width: 2.1rem;
-  height: 1.7rem;
-  cursor: pointer;
-}
-
-ul.popup {
-  padding: 0;
-  margin: 0;
-  width: 10rem;
-  list-style: none;
-  background-color: #fff;
-  position: absolute;
-  top: 1rem;
-  left: 1rem;
-  border: 1px solid #ebebeb;
-  border-radius: 5px 0 5px 5px;
-}
-
-ul.popup li {
-  padding: 12px 7px;
-  text-align: center;
-}
-
-ul.popup li.delete,
-ul.popup li.add-in-folder {
-  font-size: 1.4rem;
-  font-weight: 400;
-  color: var(--gray-100-color);
-}
-
-ul.popup li:hover {
-  background-color: var(--gray-10-color);
-  color: var(--primary-color);
-  cursor: pointer;
-}
