--- src/components/Components/Option.css
@@ -1,16 +1,0 @@
-.option {
-  display: flex;
-  gap: 1.2rem;
-}
-.option-button {
-  display: flex;
-  align-items: center;
-  gap: 0.4rem;
-}
-
-.text-style {
-  display: flex;
-  font-size: 1.4rem;
-  color: var(--gray-60-color);
-  align-items: center;
-}
