--- src/components/Components/Profile.css
@@ -1,19 +1,0 @@
-.profile {
-  background-color: var(--gray-bg-color);
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-top: 2rem;
-  padding-bottom: 6rem;
-  gap: 2rem;
-}
-
-.profile-img {
-  width: 6rem;
-  height: 6rem;
-}
-
-.profile-name {
-  font-size: 16px;
-}
