--- src/components/Components/SearchBar.css
@@ -1,20 +1,0 @@
-.search-bar-container {
-  position: relative;
-}
-
-.search-bar-container > img {
-  position: absolute;
-  top: 50%;
-  left: 1.6rem;
-  transform: translateY(-50%);
-}
-
-.searchBar {
-  background-color: #f5f5f5;
-  width: 100%;
-  height: 5.4rem;
-  border: none;
-  border-radius: 1rem;
-  padding-left: 4rem;
-  margin: 4rem auto;
-}
