--- src/pages/Folder/FolderPage.css
@@ -1,13 +1,0 @@
-.folder-section {
-  max-width: 106rem;
-  margin: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding-top: 4rem;
-  padding-bottom: 4rem;
-
-  @media (max-width: 1200px) {
-    padding: 0 3.2rem;
-  }
-}
