--- src/pages/Shared/SharedPage.css
@@ -1,19 +1,0 @@
-folder {
-  background-color: var(--white-color);
-  padding-top: 4rem;
-  padding-bottom: 4rem;
-}
-
-.folder {
-  max-width: 106rem;
-  margin: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding-top: 4rem;
-  padding-bottom: 4rem;
-
-  @media (max-width: 1200px) {
-    padding: 0 3.2rem;
-  }
-}
